put fns for creating activities in one module, use consistent names
create_curation_activity is not a template helper
