+++ ckanext/unhcr/activity.py
@@ -2,8 +2,9 @@ from ckan import model
 import ckan.plugins.toolkit as toolkit
 
 
-def log_download_activity(context, resource_id):
-    """Log a resource download activity in the activity stream
+def create_download_activity(context, resource_id):
+    """
+    Log a 'resource download' activity in the activity stream
     """
     user = context['user']
     user_id = None
@@ -29,3 +30,28 @@ def log_download_activity(context, resource_id):
 
     create_activity = toolkit.get_action('activity_create')
     create_activity(activity_create_context, activity_dict)
+
+
+def create_curation_activity(
+        activity_type, dataset_id, dataset_name, user_id,
+        message=None, **kwargs):
+    """
+    Log a 'changed curation state' activity in the activity stream
+    """
+    activity_context = {'ignore_auth': True}
+    data_dict = {
+        'user_id': user_id,
+        'object_id': dataset_id,
+        'activity_type': 'changed curation state',
+        'data': {
+            'curation_activity': activity_type,
+            'package': {'name': dataset_name, 'id': dataset_id},
+        }
+    }
+    if message:
+        data_dict['data']['message'] = message
+    if kwargs:
+        for key, value in kwargs.iteritems():
+            data_dict['data'][key] = value
+
+    toolkit.get_action('activity_create')(activity_context, data_dict)
